pyproject.toml
[tool.poetry]
name = "fastapiclinic"
version = "0.1.0"
description = ""
authors = ["Vlad Onishcuk <vlad.onischuk1234@gmail.com>"]
package-mode  = false

[tool.poetry.dependencies]
python = "3.12.7"
annotated-types = "0.7.0"
anyio = "4.6.2.post1"
click = "8.1.7"
fastapi = "0.115.2"
h11 = "0.14.0"
httptools = "0.6.4"
idna = "3.10"
pydantic = "2.9.2"
pydantic-core = "2.23.4"
python-dotenv = "1.0.1"
pyyaml = "6.0.2"
sniffio = "1.3.1"
starlette = "0.40.0"
typing-extensions = "4.12.2"
uvicorn = "0.32.0"
uvloop = "0.21.0"
watchfiles = "0.24.0"
websockets = "13.1"
fastapi-users = {extras = ["sqlalchemy"], version = "^13.0.0"}
asyncpg = "^0.29.0"
python-decouple = "^3.8"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
